requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai>=1.0.0
pydantic==2.10.6
pydantic_core==2.27.2
requests>=2.31.0
sniffio==1.3.1
tqdm>=4.65.0
typing-extensions>=4.5.0
urllib3==2.3.0

# Environment variables
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Additional dependencies
colorama>=0.4.6
rich>=13.0.0
